fix: resolve Next.js route export build errors
- Moved helper functions out of API route files to separate libs
- Created consent-helpers.ts for checkRequiredConsents function
- Created deletion-helpers.ts for verifyDeletion function
- Fixed dynamic route params type for Next.js 15 (Promise<{id: string}>)
- Moved GET handler to proper dynamic route [id]/route.ts

These changes ensure only valid HTTP method handlers are exported from API routes.
